+++ CHANGES.rst
@@ -8,6 +8,8 @@ Unreleased
 -   Drop support for Python 2.7 and 3.5.
 -   Bump MarkupSafe dependency to >=1.1.
 -   Bump Babel optional dependency to >=2.1.
+-   Remove code that was marked deprecated.
+-
 
 
 2.11.2